+++ Training2/src/components/Modal.tsx
@@ -47,15 +47,15 @@ function Modal({
   other?: React.DetailedHTMLProps<React.HTMLAttributes<HTMLDivElement>, HTMLDivElement>;
 }) {
   if (!isOpen) return undefined;
-  function handelClose() {
+  function handleClose() {
     setOpen(false);
   }
   return (
     <div>
-      <div style={{ ...styleModal }} {...other} className='1' onClick={handelClose}></div>
+      <div style={{ ...styleModal }} {...other} className='1' onClick={handleClose}></div>
       <div style={{ ...styleContainer, ...sx }} className='2'>
         {children}
-        <button onClick={handelClose} hidden={isHaveButton}>
+        <button onClick={handleClose} hidden={isHaveButton}>
           Close
         </button>
       </div>
