+++ Training4/src/app/api.ts
@@ -1,12 +1,26 @@
 import countryService from './country_server';
+import jokeService from './joke_server';
 
-interface IDataCountry {
+interface IName {
+  common: string;
+  official: string;
+}
+export interface IDataCountry {
+  name: IName;
+  region: string;
+}
+
+export function fetchCountry(value = '') {
+  if (!value) return countryService.get<IDataCountry[]>('/all');
+  return countryService.get<IDataCountry[]>(`/name/${value}`);
+}
+
+interface IDataJoke {
   type: string;
   setup: string;
   punchline: string;
 }
 
-export function fetchCountry(value = '') {
-  if (!value) return countryService.get<IDataCountry>('/all');
-  return countryService.get<IDataCountry[]>(`/name/${value}`);
+export function fetchJoker() {
+  return jokeService.get<IDataJoke>('/random');
 }
